+++ modules/mpi/ospray/MPIDistributedDevice.h
@@ -1,4 +1,4 @@
-// Copyright 2009-2021 Intel Corporation
+// Copyright 2009-2022 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
@@ -152,6 +152,8 @@ struct MPIDistributedDevice : public api::Device
       OSPFrameBuffer, OSPRenderer, OSPCamera, OSPWorld, const vec2f &) override;
 
  private:
+  ObjectHandle allocateHandle();
+
   bool initialized{false};
   bool shouldFinalizeMPI{false};