+++ modules/mpi/ospray/fb/DistributedFrameBuffer.cpp
@@ -37,7 +37,8 @@ void DistributedTileError::sync()
   if (tiles <= 0)
     return;
 
-  mpicommon::bcast(tileErrorBuffer.data(), tiles, MPI_FLOAT, 0, group.comm).wait();
+  mpicommon::bcast(tileErrorBuffer.data(), tiles, MPI_FLOAT, 0, group.comm)
+      .wait();
 }
 
 // DistributedFrameBuffer definitions ///////////////////////////////////////